Removed Spelling Mistake and Refactored filename apiFunctions.dart
