+++ lib/views/pages/events/addTaskDialog.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 
 //pages are called here
 import 'package:talawa/services/Queries.dart';
-import 'package:talawa/utils/apiFuctions.dart';
+import 'package:talawa/utils/apiFunctions.dart';
 import 'package:intl/intl.dart';
 import 'package:talawa/utils/uidata.dart';
