+++ lib/views/pages/events/taskList.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 //pages are imported here
 import 'package:talawa/services/Queries.dart';
 import 'package:talawa/services/preferences.dart';
-import 'package:talawa/utils/apiFuctions.dart';
+import 'package:talawa/utils/apiFunctions.dart';
 
 // ignore: must_be_immutable
 class TaskList extends StatefulWidget {
