+++ lib/views/pages/home_page.dart
@@ -12,7 +12,7 @@ import 'package:talawa/views/pages/members/members.dart';
 
 import 'package:talawa/views/pages/events/events.dart';
 import 'package:talawa/views/pages/chat/groups.dart';
-import 'package:talawa/utils/apiFuctions.dart';
+import 'package:talawa/utils/apiFunctions.dart';
 import 'package:persistent_bottom_nav_bar/persistent-tab-view.dart';
 import 'organization/profile_page.dart';
 import 'package:talawa/services/preferences.dart';
