+++ lib/views/pages/members/RegEventstab.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 //packages for pages are imported here
 import 'package:talawa/services/Queries.dart';
 import 'package:talawa/services/preferences.dart';
-import 'package:talawa/utils/apiFuctions.dart';
+import 'package:talawa/utils/apiFunctions.dart';
 
 
 // ignore: must_be_immutable
